+++ src/test/java/lotto/view/InputValidatorTest.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.view;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
